Add azure storage support for LangChain 1.0 (#182)
Changes include:
* Updating package version to 1.0 to mirror LangChain core bump to 1.0
* Remove support for Python 3.9 since LangChain 1.0 drops support
* Update dependencies to rely on langchain-core 1.0 release. Also, bulk
updated dependency locks to make sure we are running CI on latest
version of dependencies.
* Update package classifier
* Add beta decorator to document loader. This will allow us to get
feedback on the interface before marking it as stable.

The beta warning message looks like this when the loader is
instantiated:
```
C:\Users\kyleknapp\GitHub\langchain-azure\libs\azure-storage\demos\app.py:14: LangChainBetaWarning: `AzureBlobStorageLoader` is in public preview. Its API is not stable and may change in future versions.
  loader = AzureBlobStorageLoader(
```