Update CI to test on 3.10 (#225)
As part of LangChain 1.0, Python 3.9 is being dropped. This commit
updates the CI so that azure langchain packages that update to drop 3.9
as well will be testing under the correct Python version lower bound
(instead of GitHub possibly falling back to a different Python version
that is available in the environment).

Only difference with this PR:
https://github.com/langchain-ai/langchain-azure/pull/183 is this PR's
source branch lives on the main repository instead of a fork to see if
repository state/permissions are impacting the postgres CI.